+++ Editor/PlasticSCM/AssetOverlays/Cache/LockStatusCache.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 
 using Codice;
-using Codice.Client.BaseCommands;
 using Codice.Client.Commands.WkTree;
 using Codice.Client.Common;
 using Codice.Client.Common.Locks;
@@ -10,6 +9,7 @@ using Codice.Client.Common.Threading;
 using Codice.Client.Common.WkTree;
 using Codice.CM.Common;
 using Codice.Utils;
+using PlasticGui.WorkspaceWindow;
 
 namespace Unity.PlasticSCM.Editor.AssetsOverlays.Cache
 {
@@ -198,22 +198,22 @@ namespace Unity.PlasticSCM.Editor.AssetsOverlays.Cache
                 LockInfo lockInfo,
                 LockOwnerNameResolver nameResolver)
             {
-                AssetStatus status = AssetStatus.None;
-                // Filter the locks that are Retained since the plugin doesn't support them yet.
-                if (lockInfo.Status == LockInfo.LockStatus.Retained)
-                {
-                    status = AssetStatus.LockedRetained;
-                }
-                else
-                {
-                    status = CheckWorkspaceTreeNodeStatus.IsCheckedOut(node) ?
-                        AssetStatus.Locked : AssetStatus.LockedRemote;
-                }
-
                 return new LockStatusData(
-                    status,
+                    GetAssetStatus(node, lockInfo),
                     nameResolver.GetSeidName(lockInfo.SEIDData),
-                    LockWkInfo.GetWkCleanName(lockInfo));
+                    BranchInfoCache.GetProtectedBranchName(
+                        node.RepSpec, lockInfo.HolderBranchId));
+            }
+
+            static AssetStatus GetAssetStatus(
+                WorkspaceTreeNode node,
+                LockInfo lockInfo)
+            {
+                if (lockInfo.Status == LockInfo.LockStatus.Retained)
+                    return AssetStatus.Retained;
+
+                return CheckWorkspaceTreeNodeStatus.IsCheckedOut(node) ?
+                    AssetStatus.Locked : AssetStatus.LockedRemote;
             }
         }